refactor(structure): group queue sources
