+++ CMakeLists.txt
@@ -65,6 +65,11 @@ file(GLOB ${PROJECT_NAME}_STRUCTURE_SOURCES CONFIGURE_DEPENDS
 )
 list(SORT ${PROJECT_NAME}_STRUCTURE_SOURCES)
 
+file(GLOB ${PROJECT_NAME}_STRUCTURE_QUEUE_SOURCES CONFIGURE_DEPENDS
+     "${CMAKE_CURRENT_SOURCE_DIR}/src/structure/queue/*.cpp"
+)
+list(SORT ${PROJECT_NAME}_STRUCTURE_QUEUE_SOURCES)
+
 file(GLOB ${PROJECT_NAME}_MIDDLEWARE_SOURCES CONFIGURE_DEPENDS
      "${CMAKE_CURRENT_SOURCE_DIR}/src/middleware/*.cpp"
 )
@@ -89,6 +94,7 @@ target_sources(
   PRIVATE ${${PROJECT_NAME}_SYSTEM_EXT_SOURCES}
   PRIVATE ${${PROJECT_NAME}_UTILS_SOURCES}
   PRIVATE ${${PROJECT_NAME}_STRUCTURE_SOURCES}
+  PRIVATE ${${PROJECT_NAME}_STRUCTURE_QUEUE_SOURCES}
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_SOURCES}
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_SUBDIR_SOURCES}
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_NESTED_SUBDIR_SOURCES}
@@ -142,6 +148,7 @@ set(_xr_public_include_dirs
     ${CMAKE_CURRENT_SOURCE_DIR}/src/system
     ${CMAKE_CURRENT_SOURCE_DIR}/src/utils
     ${CMAKE_CURRENT_SOURCE_DIR}/src/structure
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/structure/queue
     ${CMAKE_CURRENT_SOURCE_DIR}/src/middleware
     ${CMAKE_CURRENT_SOURCE_DIR}/src/driver
     ${CMAKE_CURRENT_SOURCE_DIR}/system/${_xr_system}
